--- src/main/scala/charlauma/etlscala/Data.scala
@@ -1,3 +1,0 @@
-package charlauma.etlscala
-
-case class Data(id: Int, value: String)
