+++ src/main/scala/uma/charla/etl/SData.scala
@@ -1,0 +1,3 @@
+package uma.charla.etl
+
+case class SData(id: Int, value: String)
